# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chemicals API - Week 04 - Auth - 2 Collections",
    "version": "1.0.0",
    "description": "API with Chemicals and Suppliers collections - Week 4 adds GitHub OAuth. GET is public, POST/PUT/DELETE requires login."
  },
  "host": "chemicals-kw61.onrender.com",
  "basePath": "/",
  "schemes": ["https"],
  "tags": [
    { "name": "Chemicals", "description": "Chemicals collection" },
    { "name": "Suppliers", "description": "Suppliers collection - 2nd collection" },
    { "name": "Authentication", "description": "OAuth Login/Logout" }
  ],
  "securityDefinitions": {
    "githubOAuth": {
      "type": "oauth2",
      "authorizationUrl": "https://github.com/login/oauth/authorize",
      "tokenUrl": "https://github.com/login/oauth/access_token",
      "flow": "accessCode",
      "scopes": {
        "read:user": "Read user profile"
      }
    }
  },
  "definitions": {
    "Chemical": {
      "type": "object",
      "required": ["name", "concentration", "batch_no", "supplier", "mfg_date", "expiry_date", "quantity", "unit"],
      "properties": {
        "name": { "type": "string", "example": "Sulfuric Acid" },
        "concentration": { "type": "string", "example": "98%" },
        "batch_no": { "type": "string", "example": "SA-2026-001" },
        "supplier": { "type": "string", "example": "ChemSupply Ltd" },
        "mfg_date": { "type": "string", "format": "date", "example": "2026-01-15" },
        "expiry_date": { "type": "string", "format": "date", "example": "2027-01-15" },
        "quantity": { "type": "number", "example": 500 },
        "unit": { "type": "string", "example": "mL" },
        "location": { "type": "string", "example": "Shelf A1" },
        "hazard_class": { "type": "string", "example": "Corrosive" },
        "ppe": { "type": "array", "items": { "type": "string" }, "example": ["Gloves", "Goggles"] },
        "status": { "type": "string", "enum": ["In Stock", "Low", "Empty", "Expired"], "example": "In Stock" },
        "incompatible_with": { "type": "array", "items": { "type": "string" }, "example": ["Water", "Metals"] },
        "sds_url": { "type": "string" }
      }
    },
    "Supplier": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "name": { "type": "string", "example": "ChemSupply Ltd" },
        "contact_person": { "type": "string", "example": "John Doe" },
        "email": { "type": "string", "example": "sales@chemsupply.com" },
        "phone": { "type": "string", "example": "+1-555-0100" },
        "address": { "type": "string", "example": "123 Lab Street" }
      }
    }
  },
  "paths": {
    "/login": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Login with GitHub",
        "description": "Redirects to GitHub for OAuth login",
        "responses": { "302": { "description": "Redirect to GitHub" } }
      }
    },
    "/logout": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Logout",
        "description": "Logs out the current user",
        "responses": { "200": { "description": "Logged out" } }
      }
    },
    "/api/chemicals": {
      "get": { "tags": ["Chemicals"], "summary": "Get all chemicals", "responses": { "200": { "description": "OK" } } },
      "post": {
        "tags": ["Chemicals"],
        "summary": "Create chemical - PROTECTED",
        "description": "Requires login. Will return 401 if not authenticated.",
        "security": [{ "githubOAuth": ["read:user"] }],
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/Chemical" } }],
        "responses": { "201": { "description": "Created" }, "401": { "description": "Unauthorized - Please login" }, "400": { "description": "Validation error" } }
      }
    },
    "/api/chemicals/{id}": {
      "get": { "tags": ["Chemicals"], "summary": "Get by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }], "responses": { "200": { "description": "OK" } } },
      "put": {
        "tags": ["Chemicals"],
        "summary": "Update chemical - PROTECTED",
        "security": [{ "githubOAuth": ["read:user"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }, { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/Chemical" } }],
        "responses": { "200": { "description": "Updated" }, "401": { "description": "Unauthorized - Please login" } }
      },
      "delete": {
        "tags": ["Chemicals"],
        "summary": "Delete chemical - PROTECTED",
        "security": [{ "githubOAuth": ["read:user"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Deleted" }, "401": { "description": "Unauthorized - Please login" } }
      }
    },
    "/api/suppliers": {
      "get": { "tags": ["Suppliers"], "summary": "Get all suppliers - 2nd Collection", "responses": { "200": { "description": "OK" } } },
      "post": {
        "tags": ["Suppliers"],
        "summary": "Create supplier - PROTECTED",
        "security": [{ "githubOAuth": ["read:user"] }],
        "parameters": [{ "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/Supplier" } }],
        "responses": { "201": { "description": "Created" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/api/suppliers/{id}": {
      "get": { "tags": ["Suppliers"], "summary": "Get supplier by ID", "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }], "responses": { "200": { "description": "OK" } } },
      "put": {
        "tags": ["Suppliers"],
        "summary": "Update supplier - PROTECTED",
        "security": [{ "githubOAuth": ["read:user"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }, { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/Supplier" } }],
        "responses": { "200": { "description": "Updated" }, "401": { "description": "Unauthorized" } }
      },
      "delete": {
        "tags": ["Suppliers"],
        "summary": "Delete supplier - PROTECTED",
        "security": [{ "githubOAuth": ["read:user"] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "string" }],
        "responses": { "200": { "description": "Deleted" }, "401": { "description": "Unauthorized" } }
      }
    }
  }
}